实现贪吃蛇的基本功能，后续会对功能及界面进行完善

+++ Snake/CellPoint.m
@@ -34,7 +34,7 @@
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"{%d,%d}",self.x,self.y];
+    return [NSString stringWithFormat:@"{%ld,%ld}",(long)self.x,(long)self.y];
 }
 
 @end
